# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Películas & Series Latino",
    "version": "1.0.0",
    "description": "API REST en vivo para consultar catálogos estilo Netflix, metadatos y reproductores embeds verificados de películas, series, animes y doramas en audio Latino."
  },
  "servers": [
    {
      "url": "https://api-pelis-series-latino.vercel.app",
      "description": "Servidor de Producción Vercel"
    }
  ],
  "paths": {
...
    "/api/v1/home": {
      "get": {
        "summary": "Obtener el catálogo de inicio estilo Netflix (alias de /api/v1/feeds/home)",
        "description": "Mismo contrato y mismos parámetros que /api/v1/feeds/home.",
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "CL"
            }
          },
          {
            "name": "detail",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "card",
                "compact"
              ],
              "default": "card"
            }
          },
          {
            "name": "rows",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 40
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feed obtenido con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeFeedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/feeds/home": {
      "get": {
        "summary": "Obtener el catálogo de inicio estilo Netflix",
        "description": "Devuelve el hero rotatorio (`spotlight`, hasta 5 títulos con ficha completa) y alrededor de 15 carruseles temáticos (Top 10 del país, novedades, películas, series, mejor valoradas, anime, acción, comedia, terror, ciencia ficción, animación, drama, romance, aventura, crimen, documentales y clásicos). Cada fila se omite si no reúne al menos 8 títulos, de modo que nunca se publica un carrusel a medias.\n\nLas tarjetas viajan con la metadata completa (sinopsis, tagline, logo, duración, clasificación por edades, géneros, tráiler y número de temporadas), por lo que un cliente puede abrir su ficha emergente SIN ninguna petición adicional; solo los enlaces de reproducción se resuelven aparte en `/api/v1/media/{id}/streams`.",
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "required": false,
            "description": "Código ISO del país, usado en el titular del Top 10.",
            "schema": {
              "type": "string",
              "default": "CL"
            }
          },
          {
            "name": "detail",
            "in": "query",
            "required": false,
            "description": "card (por defecto) → tarjetas con metadata completa. compact → proyección compacta clásica, payload mínimo.",
            "schema": {
              "type": "string",
              "enum": [
                "card",
                "compact"
              ],
              "default": "card"
            }
          },
          {
            "name": "rows",
            "in": "query",
            "required": false,
            "description": "Lista separada por comas de ids de fila (ej. action,anime,top_10_cl). Devuelve solo esas filas, para cargar el home por lotes.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Ítems por carrusel. Máximo 40.",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 40
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feed obtenido con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeFeedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/movies": {
      "get": {
        "summary": "Obtener catálogo de películas con paginación",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "Número de página real sobre el índice general, empezando en 1."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "description": "Cantidad de ítems por página. Alias compatible: size. Máximo 100."
          },
          {
            "name": "size",
            "in": "query",
            "description": "Alias de limit para definir la cantidad de ítems por página. Máximo 100.",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "genre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de películas obtenida con éxito"
          }
        }
      }
    },
    "/api/v1/series": {
      "get": {
        "summary": "Obtener catálogo de series con paginación",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "Número de página real sobre el índice general, empezando en 1."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "description": "Cantidad de ítems por página. Alias compatible: size. Máximo 100."
          },
          {
            "name": "size",
            "in": "query",
            "description": "Alias de limit para definir la cantidad de ítems por página. Máximo 100.",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "genre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de series obtenida con éxito"
          }
        }
      }
    },
    "/api/v1/media": {
      "get": {
        "summary": "Obtener catálogo general (películas + series)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "Número de página real sobre el índice general, empezando en 1."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "description": "Cantidad de ítems por página. Alias compatible: size. Máximo 100."
          },
          {
            "name": "size",
            "in": "query",
            "description": "Alias de limit para definir la cantidad de ítems por página. Máximo 100.",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Catálogo devuelto con éxito"
          }
        }
      }
    },
    "/api/v1/media/batch": {
      "get": {
        "summary": "Consultar múltiples contenidos en lote por IDs",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "description": "Lista de IDs separados por comas (ej. ?ids=558,94997,748783)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "compact",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lote de contenidos devuelto con éxito"
          }
        }
      }
    },
...
    "/api/v1/discover": {
      "get": {
        "summary": "Descubrimiento paginado del catálogo general",
        "description": "Devuelve una página real del índice general y permite filtrar por type o genre.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número de página real sobre el índice general, empezando en 1.",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Cantidad de ítems por página. Alias compatible: size. Máximo 100.",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Alias de limit para definir la cantidad de ítems por página. Máximo 100.",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Descubrimiento devuelto con éxito"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CardItem": {
        "type": "object",
        "description": "Tarjeta de carrusel del home (detail=card). Lleva toda la metadata necesaria para pintar una ficha emergente sin peticiones adicionales; no incluye reparto, servidores ni temporadas.",
        "properties": {
          "id": {
            "type": "string"
          },
          "tmdb_id": {
            "type": "integer"
          },
          "type": {
            "type": "string",
            "enum": [
              "movie",
              "tvseries"
            ]
          },
          "title": {
            "type": "string"
          },
          "original_title": {
            "type": "string"
          },
          "tagline": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "poster": {
            "type": "string",
            "nullable": true
          },
          "backdrop": {
            "type": "string",
            "nullable": true
          },
          "logo": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "number"
          },
          "content_rating": {
            "type": "string",
            "nullable": true
          },
          "runtime": {
            "type": "integer",
            "nullable": true
          },
          "release_date": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "nullable": true
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subcategories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "trailer": {
            "type": "string",
            "nullable": true
          },
          "total_seasons": {
            "type": "integer"
          },
          "total_episodes": {
            "type": "integer"
          },
          "detail_url": {
            "type": "string",
            "example": "/api/v1/media/scary-movie-6"
          },
          "streams_url": {
            "type": "string",
            "example": "/api/v1/media/scary-movie-6/streams"
          }
        }
      },
      "HeroItem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CardItem"
          },
          {
            "type": "object",
            "description": "Destacado del hero: la tarjeta más reparto y dirección.",
            "properties": {
              "cast": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "cast_details": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "dubbing_cast": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "director": {
                "type": "string",
                "nullable": true
              },
              "created_by": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "HomeFeedRow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "top_10_cl"
          },
          "title": {
            "type": "string",
            "example": "Top 10 en Chile hoy"
          },
          "subtitle": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "carousel",
              "featured"
            ]
          },
          "layout": {
            "type": "string",
            "enum": [
              "ranked",
              "backdrop",
              "poster"
            ],
            "description": "Pista de render: ranked (numerado tipo Top 10), backdrop (tarjeta ancha) o poster (tarjeta vertical)."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardItem"
            }
          },
          "endpoint": {
            "type": "string",
            "description": "Endpoint paginado para el 'ver todo' / scroll infinito de la fila.",
            "example": "/api/v1/discover?genre=Acci%C3%B3n"
          }
        }
      },
      "HomeFeedResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "featured": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/HeroItem"
                  }
                ],
                "nullable": true,
                "description": "Primer destacado. Se mantiene por compatibilidad; equivale a spotlight[0]."
              },
              "spotlight": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/HeroItem"
                },
                "description": "Hasta 5 destacados con ficha completa para rotar en el hero."
              },
              "rows": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/HomeFeedRow"
                }
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
...
    }
  }
}